Move class_identifier update to zero initializer
Avoids having to follow a two-step process (with the risk of forgetting the
second step).
